+++ pkg/coderefs/alias.go
@@ -1,15 +1,33 @@
 package coderefs
 
 import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"os/exec"
+	"path/filepath"
+	"regexp"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/iancoleman/strcase"
 	"github.com/launchdarkly/ld-find-code-refs/internal/helpers"
 	"github.com/launchdarkly/ld-find-code-refs/internal/options"
+	"github.com/launchdarkly/ld-find-code-refs/internal/validation"
 )
 
-func generateAliases(flags []string, aliases []options.Alias) (map[string][]string, error) {
+func generateAliases(flags []string, aliases []options.Alias, dir string) (map[string][]string, error) {
+	allFileContents, err := processFileContent(aliases, dir)
+	if err != nil {
+		return nil, err
+	}
+
 	ret := make(map[string][]string, len(flags))
 	for _, flag := range flags {
 		for _, a := range aliases {
-			flagAliases, err := a.Generate(flag)
+			flagAliases, err := generateAlias(a, flag, dir, allFileContents)
 			if err != nil {
 				return nil, err
 			}
@@ -18,4 +36,112 @@ func generateAliases(flags []string, aliases []options.Alias) (map[string][]stri
 		ret[flag] = helpers.Dedupe(ret[flag])
 	}
 	return ret, nil
+}
+
+func generateAlias(a options.Alias, flag, dir string, allFileContents map[string][]byte) ([]string, error) {
+	ret := []string{}
+	switch a.Type.Canonical() {
+	case options.Literal:
+		ret = a.Flags[flag]
+	case options.CamelCase:
+		ret = []string{strcase.ToLowerCamel(flag)}
+	case options.PascalCase:
+		ret = []string{strcase.ToCamel(flag)}
+	case options.SnakeCase:
+		ret = []string{strcase.ToSnake(flag)}
+	case options.UpperSnakeCase:
+		ret = []string{strcase.ToScreamingSnake(flag)}
+	case options.KebabCase:
+		ret = []string{strcase.ToKebab(flag)}
+	case options.DotCase:
+		ret = []string{strcase.ToDelimited(flag, '.')}
+	case options.FilePattern:
+		// Concatenate the contents of all files into a single byte array to be matched by specified patterns
+		fileContents := []byte{}
+		for _, path := range a.Paths {
+			pathFileContents := allFileContents[path]
+			fileContents = append(fileContents, pathFileContents...)
+		}
+
+		for _, p := range a.Patterns {
+			pattern := regexp.MustCompile(strings.ReplaceAll(p, "FLAG_KEY", flag))
+			results := pattern.FindAllStringSubmatch(string(fileContents), -1)
+			for _, res := range results {
+				if len(res) > 1 {
+					ret = append(ret, res[1:]...)
+				}
+			}
+		}
+	case options.Command:
+		ctx := context.Background()
+		if a.Timeout != nil && *a.Timeout > 0 {
+			var cancel context.CancelFunc
+			ctx, cancel = context.WithDeadline(ctx, time.Now().Add(time.Second*time.Duration(*a.Timeout)))
+			defer cancel()
+		}
+		tokens := strings.Split(*a.Command, " ")
+		name := tokens[0]
+		args := []string{}
+		if len(tokens) > 1 {
+			args = tokens[1:]
+		}
+		/* #nosec */
+		cmd := exec.CommandContext(ctx, name, args...)
+		cmd.Stdin = strings.NewReader(flag)
+		cmd.Dir = dir
+		stdout, err := cmd.Output()
+		if err != nil {
+			return nil, fmt.Errorf("failed to execute alias command: %w", err)
+		}
+		err = json.Unmarshal(stdout, &ret)
+		if err != nil {
+			return nil, fmt.Errorf("could not unmarshal json output of alias command: %w", err)
+		}
+	}
+
+	return ret, nil
+}
+
+// processFileContent reads and stores the content of files specified by filePattern alias matchers to be matched for aliases
+func processFileContent(aliases []options.Alias, dir string) (map[string][]byte, error) {
+	allFileContents := map[string][]byte{}
+	for idx, a := range aliases {
+		if a.Type != options.FilePattern {
+			continue
+		}
+
+		aliasId := strconv.Itoa(idx)
+		if a.Name != "" {
+			aliasId = a.Name
+		}
+
+		paths := []string{}
+		for _, glob := range a.Paths {
+			absGlob := filepath.Join(dir, glob)
+			matches, err := filepath.Glob(absGlob)
+			if err != nil {
+				return nil, fmt.Errorf("filepattern '%s': could not process path glob '%s'", aliasId, absGlob)
+			}
+			paths = append(paths, matches...)
+		}
+		paths = helpers.Dedupe(paths)
+
+		for _, path := range paths {
+			_, pathAlreadyProcessed := allFileContents[path]
+			if pathAlreadyProcessed {
+				continue
+			}
+
+			if !validation.FileExists(path) {
+				return nil, fmt.Errorf("filepattern '%s': could not find file at path '%s'", aliasId, path)
+			}
+			/* #nosec */
+			data, err := ioutil.ReadFile(path)
+			if err != nil {
+				return nil, fmt.Errorf("filepattern '%s': could not process file at path '%s': %v", aliasId, path, err)
+			}
+			allFileContents[path] = data
+		}
+	}
+	return allFileContents, nil
 }